server_code/requirements.txt
google-genai==1.43.0
google-generativeai==0.8.5